fix: move config out from pkg directory
